+++ src/libstd/thunk.rs
