+++ src/test/run-pass/match-ref-binding-in-guard-3256.rs
@@ -8,11 +8,11 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
-extern crate rustrt;
+use std::rt::exclusive;
 
 pub fn main() {
     unsafe {
-        let x = Some(::rustrt::exclusive::Exclusive::new(true));
+        let x = Some(exclusive::Exclusive::new(true));
         match x {
             Some(ref z) if *z.lock() => {
                 assert!(*z.lock());
